% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG_from_Ordering.R
\name{DAG_from_Ordering}
\alias{DAG_from_Ordering}
\title{Infer  DAG using topological ordering}
\usage{
DAG_from_Ordering(
  X,
  TO,
  mtd = "ztest",
  alpha = 0.05,
  threshold = 0.1,
  FCD = NULL,
  precmtd = NULL
)
}
\arguments{
\item{X:}{data in n x p matrix}

\item{TO:}{topological ordering}

\item{alpha:}{desired selection significance level}

\item{mtd:}{methods for learning DAG from topological orderings.
"ztest": (p<n) [Multiple Testing and Error Control in Gaussian Graphical Model Selection. Drton and Perlman.2007]
"rls": (p<n) fit recursive least squares using ggm package and threshold the regression coefs
"chol": (p<n) perform cholesky decomposition and threshold the regression coefs
"dlasso": debiased lasso (default with FCD=True and precmtd="sqrtlasso");
 "lasso": lasso with fixed lambda from [Penalized likelihood methods for estimation of sparse high-dimensional directed acyclic graphs. Shojaie and Michailidis. 2010];
 "adalasso": adaptive lasso with fixed lambda from [Shojaie and Michailidis. 2010];
 "cvlasso": cross-validated lasso from glmnet;
  "scallasso": scaled lasso.}

\item{threshold:}{only used in rls and chol. the hard threshold level.}

\item{FCD:}{only used in debiased lasso,  the FCD procedure [False Discovery Rate Control via Debiased Lasso. Javanmard and Montanari. 2018]
or use individual tests to select support.}

\item{precmtd:}{only used in debiased lasso, how to compute debiasing matrix
"cv": node-wise lasso w/ joint 10 fold cv
"sqrtlasso": square-root lasso (no tune, default)}
}
\value{
Adjacency matrix with ADJ[i,j]!=0 iff i->j
}
\description{
Infer  DAG using topological ordering
}
