% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqVarDAG_BU.R
\name{EqVarDAG_BU}
\alias{EqVarDAG_BU}
\title{Estimate topological ordering and DAG using Bottom-up approach}
\usage{
EqVarDAG_BU(X)
}
\arguments{
\item{X}{An n-by-p data matrix.}
}
\value{
Estimated Adjacency matrix and topological ordering.
}
\description{
Estimate topological ordering and DAG using Bottom-up approach
}
\examples{
X1<-rnorm(100)
X2<-X1+rnorm(100)
EqVarDAG_BU(cbind(X1,X2))

#$adj
#[,1] [,2]
#[1,]    0    1
#[2,]    0    0
#
#$TO
#[1] 1 2
}
