% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqVarDAG_HD_CLIME.R
\name{EqVarDAG_HD_CLIME}
\alias{EqVarDAG_HD_CLIME}
\title{Estimate topological ordering and DAG using high dimensional bottom-up CLIME approach}
\usage{
EqVarDAG_HD_CLIME(X, cv = T, lambdafix = 0.1)
}
\arguments{
\item{X}{An n-by-p data matrix.}

\item{cv}{Obtain regularization parameter by cross-validation or use speficied value. Default True.}

\item{lambdafix}{Specified lambda. (If cv=F).}
}
\value{
Estimated Adjacency matrix and topological ordering.
}
\description{
Estimate topological ordering and DAG using high dimensional bottom-up CLIME approach
}
\examples{
X1<-rnorm(100)
X2<-X1+rnorm(100)
EqVarDAG_HD_TD(cbind(X1,X2),2)

#$adj
#[,1] [,2]
#[1,]    0    1
#[2,]    0    0
#
#$TO
#[1] 1 2
}
