% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqVarDAG_HD_TD.R
\name{EqVarDAG_HD_TD}
\alias{EqVarDAG_HD_TD}
\title{Estimate topological ordering and DAG using high dimensional top-down approach}
\usage{
EqVarDAG_HD_TD(X, J)
}
\arguments{
\item{X}{An n-by-p data matrix.}

\item{J}{maximum number of parents to condition on.}
}
\value{
Estimated Adjacency matrix and topological ordering.
}
\description{
Estimate topological ordering and DAG using high dimensional top-down approach
}
\examples{
X1<-rnorm(100)
X2<-X1+rnorm(100)
EqVarDAG_HD_TD(cbind(X1,X2),2)

#$adj
#[,1] [,2]
#[1,]    0    1
#[2,]    0    0
#
#$TO
#[1] 1 2
}
