% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vselect.R
\name{vselect}
\alias{vselect}
\title{Variable selection toolbox for lasso,}
\usage{
vselect(
  X,
  Y,
  alpha = 0.05,
  p_total = 0,
  selmtd = "dlasso",
  precmtd = "sqrtlasso",
  FCD = TRUE
)
}
\arguments{
\item{X, Y:}{n x p and 1 x p matrix}

\item{alpha:}{desired selection significance level}

\item{p_total:}{total number of variables, used in bonferroni correction}

\item{selmtd:}{variable selection method: avaiable methods:
"dlasso": debiased lasso (default with FCD=True and precmtd="sqrtlasso");
"lasso": lasso with fixed lambda from [Penalized likelihood methods for estimation of sparse high-dimensional directed acyclic graphs. Shojaie and Michailidis. 2010];
"adalasso": adaptive lasso with fixed lambda from [Shojaie and Michailidis. 2010];
"cvlasso": cross-validated lasso from glmnet;
 "scallasso": scaled lasso.}

\item{FCD:}{for debiased lasso, use the FCD procedure [False Discovery Rate Control via Debiased Lasso. Javanmard and Montanari. 2018]
or use individual tests to select support.}

\item{precmtd:}{for debiased lasso, how to compute debiasing matrix
"cv": node-wise lasso w/ joint 10 fold cv
"sqrtlasso": square-root lasso (no tune, default)}
}
\value{
one-hot model selection vector (1 x p) and fitted residuals
}
\description{
Variable selection toolbox for lasso,
}
