% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EqVarDAG_HD_CLIME.R
\name{EqVarDAG_HD_CLIME}
\alias{EqVarDAG_HD_CLIME}
\title{Estimate topological ordering and DAG using high dimensional bottom-up CLIME approach
Estimate  DAG using topological ordering}
\usage{
EqVarDAG_HD_CLIME(
  X,
  mtd = "dlasso",
  alpha = 0.05,
  threshold = 0.1,
  FCD = TRUE,
  precmtd = "sqrtlasso"
)
}
\arguments{
\item{X, Y:}{n x p and 1 x p matrix}

\item{alpha:}{desired selection significance level}

\item{mtd:}{methods for learning DAG from topological orderings.
"ztest": (p<n) [Multiple Testing and Error Control in Gaussian Graphical Model Selection. Drton and Perlman.2007]
"rls": (p<n) fit recursive least squares using ggm package and threshold the regression coefs
"chol": (p<n) perform cholesky decomposition and threshold the regression coefs
"dlasso": debiased lasso (default with FCD=True and precmtd="sqrtlasso");
 "lasso": lasso with fixed lambda from [Penalized likelihood methods for estimation of sparse high-dimensional directed acyclic graphs. Shojaie and Michailidis. 2010];
 "adalasso": adaptive lasso with fixed lambda from [Shojaie and Michailidis. 2010];
 "cvlasso": cross-validated lasso from glmnet;
  "scallasso": scaled lasso.}

\item{threshold:}{for rls and chol, the threshold level.}

\item{FCD:}{for debiased lasso, use the FCD procedure [False Discovery Rate Control via Debiased Lasso. Javanmard and Montanari. 2018]
or use individual tests to select support.}

\item{precmtd:}{for debiased lasso, how to compute debiasing matrix
"cv": node-wise lasso w/ joint 10 fold cv
"sqrtlasso": square-root lasso (no tune, default)}
}
\value{
Adjacency matrix with ADJ[i,j]!=0 iff i->j, and topological ordering
}
\description{
Estimate topological ordering and DAG using high dimensional bottom-up CLIME approach
Estimate  DAG using topological ordering
}
\examples{
X1<-rnorm(100)
X2<-X1+rnorm(100)
EqVarDAG_HD_TD(cbind(X1,X2),2)

#$adj
#[,1] [,2]
#[1,]    0    1
#[2,]    0    0
#
#$TO
#[1] 1 2
}
